+++ src/message.rs
@@ -2,16 +2,10 @@
 pub enum PlotUiMessage {
     ToggleLegend,
     ToggleSeriesVisibility(String),
-    // /// UI payload to render a tooltip overlay. None clears it.
-    // TooltipUiChanged(Option<TooltipUiPayload>),
-    // /// UI payload to render a small cursor-position overlay in the plot.
-    // /// None clears it.
-    // CursorPositionUiChanged(Option<CursorPositionUiPayload>),
     RenderUpdate(PlotRenderUpdate),
 }
 
 /// Context passed to a tooltip formatting callback.
-/// Public so applications can implement a custom formatter.
 #[derive(Debug, Clone)]
 pub struct TooltipContext {
     /// Label of the series, if any (empty string means none)
@@ -19,13 +13,13 @@ pub struct TooltipContext {
     /// Index within the series [0..len)
     pub point_index: usize,
     /// Data-space coordinates
-    pub x: f32,
-    pub y: f32,
+    pub x: f64,
+    pub y: f64,
 }
 
 #[derive(Debug, Clone)]
 pub struct TooltipUiPayload {
-    pub x: f32, // screen-space logical px relative to shader area
+    pub x: f32,
     pub y: f32,
     pub text: String,
 }
@@ -34,8 +28,8 @@ pub struct TooltipUiPayload {
 #[derive(Debug, Clone)]
 pub struct CursorPositionUiPayload {
     /// World/data-space coordinates for the cursor
-    pub x: f32,
-    pub y: f32,
+    pub x: f64,
+    pub y: f64,
     /// Formatted text to render
     pub text: String,
 }
