+++ src/point.rs
@@ -1,5 +1,3 @@
-use bytemuck::{Pod, Zeroable};
-
 #[repr(u32)]
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum MarkerType {
@@ -11,38 +9,38 @@ pub enum MarkerType {
 }
 
 #[repr(C)]
-#[derive(Copy, Clone, Debug, Pod, Zeroable)]
+#[derive(Copy, Clone, Debug)]
 pub struct Point {
-    pub position: [f32; 2],
+    pub position: [f64; 2],
     /// Marker size interpreted as PIXEL RADIUS (screen-space, invariant to zoom)
     pub size: f32,
 }
 
 impl Point {
-    pub fn new(x: f32, y: f32, size: f32) -> Self {
+    pub fn new(x: f64, y: f64, size: f32) -> Self {
         Self {
             position: [x, y],
             size,
         }
     }
 
-    pub fn filled_circle(x: f32, y: f32, size: f32) -> Self {
+    pub fn filled_circle(x: f64, y: f64, size: f32) -> Self {
         Self::new(x, y, size)
     }
 
-    pub fn empty_circle(x: f32, y: f32, size: f32) -> Self {
+    pub fn empty_circle(x: f64, y: f64, size: f32) -> Self {
         Self::new(x, y, size)
     }
 
-    pub fn square(x: f32, y: f32, size: f32) -> Self {
+    pub fn square(x: f64, y: f64, size: f32) -> Self {
         Self::new(x, y, size)
     }
 
-    pub fn star(x: f32, y: f32, size: f32) -> Self {
+    pub fn star(x: f64, y: f64, size: f32) -> Self {
         Self::new(x, y, size)
     }
 
-    pub fn triangle(x: f32, y: f32, size: f32) -> Self {
+    pub fn triangle(x: f64, y: f64, size: f32) -> Self {
         Self::new(x, y, size)
     }
 }
